fix: load file configuration from CLI parameter
