+++ config/commits.go
@@ -4,6 +4,14 @@ import (
 	"github.com/aevea/commitsar/internal/root_runner"
 	"github.com/aevea/integrations"
 	"github.com/spf13/viper"
+	"fmt"
+	"os"
+	"log"
+)
+
+const (
+	// CommitsarConfigPath is used an env variable to override the default location of the config file.
+	CommitsarConfigPath = "COMMITSAR_CONFIG_PATH"
 )
 
 // CommitConfig will return the RunnerOptions using defaults unless overridden in config or flags
@@ -14,6 +22,11 @@ func CommitConfig() root_runner.RunnerOptions {
 	all := false
 	upstreamBranch := integrations.FindCompareBranch()
 	requiredScopes := []string{}
+
+	if err := LoadConfig(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
 
 	if viper.IsSet("commits.strict") {
 		strict = viper.GetBool("commits.strict")
@@ -42,4 +55,28 @@ func CommitConfig() root_runner.RunnerOptions {
 		UpstreamBranch: upstreamBranch,
 		RequiredScopes: requiredScopes,
 	}
+}
+
+// LoadConfig iterates through possible config paths. No config will be loaded if no files are present.
+func LoadConfig() error {
+	viper.AutomaticEnv()
+	viper.SetConfigName(".commitsar")
+	viper.SetConfigType("yaml")
+
+	if viper.IsSet(CommitsarConfigPath) {
+		viper.AddConfigPath(viper.GetString(CommitsarConfigPath))
+	}
+
+	viper.AddConfigPath(viper.GetString("config-path"))
+
+	if err := viper.ReadInConfig(); err != nil {
+		if _, ok := err.(viper.ConfigFileNotFoundError); ok {
+			// Config file not found; ignore error if desired
+			log.Println("config file not found, using defaults")
+		} else {
+			// Config file was found but another error was produced
+			return err
+		}
+	}
+	return nil
 }
